+++ src/Binrep/Example.hs
@@ -11,7 +11,8 @@ import Binrep.Type.Int
 import GHC.Generics ( Generic )
 import Data.Data ( Data )
 
-brCfgNoSum :: BR.Cfg (I 'U 'I1 'LE)
+type BrNoSum = I 'U 'I1 'LE
+brCfgNoSum :: BR.Cfg BrNoSum
 brCfgNoSum = BR.Cfg { BR.cSumTag = undefined }
 
 data DV
@@ -27,7 +28,8 @@ instance Get  DV where get  = getGeneric  brCfgNoSum
 data DU = DU
     deriving stock (Generic, Data, Show, Eq)
 
-instance BLen DU where blen = blenGeneric brCfgNoSum
+--instance BLen DU where blen = blenGeneric brCfgNoSum
+instance BLen DU where type CBLen DU = CBLenGeneric BrNoSum DU
 instance Put  DU where put  = putGeneric  brCfgNoSum
 instance Get  DU where get  = getGeneric  brCfgNoSum
 
@@ -40,28 +42,26 @@ data DSS = DSS
   } deriving stock (Generic, Data, Show, Eq)
 
 instance BLen DSS where blen = blenGeneric brCfgNoSum
+--instance BLen DSS where type CBLen DSS = CBLenGeneric BrNoSum DSS
 instance Put  DSS where put  = putGeneric  brCfgNoSum
 instance Get  DSS where get  = getGeneric  brCfgNoSum
 
 data DCS = DCS1 {- DSS -} | DCS2 | DCS3 | DCS4 | DCS5
     deriving stock (Generic, Data, Show, Eq)
 
-brCfgDCS :: BR.Cfg (I 'U 'I1 'LE)
+type BrSumDCS = I 'U 'I1 'LE
+brCfgDCS :: BR.Cfg BrSumDCS
 brCfgDCS = BR.Cfg { BR.cSumTag = BR.cSumTagHex $ drop 3 }
 
---instance BR.BLen DCS where blen = BR.blenGeneric brCfgDCS
-deriving anyclass instance BLen DCS
+--instance BLen DCS where blen = BR.blenGeneric brCfgDCS
+instance BLen DCS where type CBLen DCS = CBLenGeneric BrSumDCS DCS
 instance Put  DCS where put  = putGeneric  brCfgDCS
 instance Get  DCS where get  = getGeneric  brCfgDCS
 
 data DX = DX DU
     deriving stock (Generic, Data, Show, Eq)
 
-type instance CBLen DX  = CBLenGeneric (I 'U 'I1 'LE) DX
-type instance CBLen DU  = CBLenGeneric (I 'U 'I1 'LE) DU
-type instance CBLen DSS = CBLenGeneric (I 'U 'I1 'LE) DSS
-type instance CBLen DCS = CBLenGeneric (I 'U 'I1 'LE) DCS
-deriving anyclass instance BLen DX
-
--- TODO clean up that CBLen messing around, probably don't mention it outside
--- the module (it's weird and will probably break things)
+--instance BLen DX where blen = blenGeneric brCfgNoSum
+instance BLen DX where type CBLen DX = CBLenGeneric BrNoSum DX
+instance Put  DX where put  = putGeneric  brCfgNoSum
+instance Get  DX where get  = getGeneric  brCfgNoSum